+++ src/Opulence/Framework/Bootstrappers/Http/Routing/RouterBootstrapper.php
@@ -9,6 +9,7 @@
 namespace Opulence\Framework\Bootstrappers\Http\Routing;
 
 use Opulence\Bootstrappers\Bootstrapper;
+use Opulence\Framework\Routing\Dispatchers\DependencyResolver;
 use Opulence\Ioc\IContainer;
 use Opulence\Routing\Dispatchers\Dispatcher;
 use Opulence\Routing\Dispatchers\IDispatcher;
@@ -97,7 +98,7 @@ class RouterBootstrapper extends Bootstrapper
      */
     protected function getRouteDispatcher(IContainer $container) : IDispatcher
     {
-        return new Dispatcher($container);
+        return new Dispatcher(new DependencyResolver($container));
     }
 
     /**