+++ src/Opulence/Routing/Dispatchers/DependencyResolutionException.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Opulence
+ *
+ * @link      https://www.opulencephp.com
+ * @copyright Copyright (C) 2016 David Young
+ * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
+ */
+namespace Opulence\Routing\Dispatchers;
+
+use Exception;
+
+/**
+ * Defines a dependency resolution exception
+ */
+class DependencyResolutionException extends Exception
+{
+    // Don't do anything
+}